+++ src/tech/particles/ParticleModifiers.ts
@@ -13,8 +13,16 @@ export default class ParticleModifiers {
         return { name: 'randomSize', value: minPercentage };
     }
 
+    public static randomSpeed(minPercentage: number): ParticleModifier {
+        return { name: 'randomSpeed', value: minPercentage };
+    }
+
     public static parallax(): ParticleModifier {
         return { name: 'parallax', value: null };
+    }
+
+    public static turn(deg: number): ParticleModifier {
+        return { name: 'turn', value: deg };
     }
 
     // Prevent initialization on a static class