+++ src/tech/particles/emmiters/VortexParticleEmitter.ts
@@ -25,6 +25,7 @@ export default class VortexParticleEmitter extends ParticleEmitter {
             vY: TSCMath.cos(dir) * this.particleSpeed,
             color: this.particleColor,
             size: this.particleSize,
+            lifetime: this.particleLifetime,
             modifiers: this.modifiers
         }));
     }
@@ -43,6 +44,9 @@ export default class VortexParticleEmitter extends ParticleEmitter {
 
         particle.x += particle.vX * dt;
         particle.y += particle.vY * dt;
+        particle.dir += this.modifiers.has('turn')
+            ? this.modifiers.get('turn') as number * dt
+            : 0;
     }
 
     constructor(options?: Partial<VortexParticleEmitterOptions>) {